Gsuite/requirements.txt
google-api-python-client
google-oauth2-tool
httplib2
pydrive